feat(aio): response to @igorminar feedback
