+++ angular.io/src/app/app.component.spec.ts
@@ -1,5 +1,3 @@
-// /* tslint:disable:no-unused-variable */
-
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { AppComponent } from './app.component';
 import { AppModule } from './app.module';
