--- angular.io/src/app/model/index.ts
@@ -1,2 +1,0 @@
-export { Doc } from './doc';
-export { DocMetadata } from './doc-metadata';
