+++ angular.io/src/app/nav-engine/doc.model.ts
@@ -1,5 +1,10 @@
-export class DocMetadata {
+export interface DocMetadata {
   id: string;    // 'home'
   title: string; // 'Home'
   url: string;   // 'assets/documents/home.html'
+}
+
+export interface Doc {
+  metadata: DocMetadata;
+  content?: string;
 }
