+++ angular.io/src/app/nav-engine/index.ts
@@ -1,18 +1,16 @@
-import { DocBuilderService } from './doc-builder.service';
 import { DocService } from './doc.service';
 import { FileService } from './file.service';
 import { NavEngine } from './nav-engine.service';
 import { NavLinkDirective } from './nav-link.directive';
 import { SiteMapService } from './sitemap.service';
 
-export { DocBuilderService } from './doc-builder.service';
+export { Doc, DocMetadata } from './doc.model';
 
 export const navDirectives = [
   NavLinkDirective
 ];
 
 export const navProviders = [
-  DocBuilderService,
   DocService,
   FileService,
   NavEngine,
