+++ angular.io/src/app/nav-engine/nav-engine.service.ts
@@ -1,13 +1,14 @@
 import { Injectable } from '@angular/core';
 
+import { Doc } from './doc.model';
 import { DocService } from './doc.service';
 
 @Injectable()
 export class NavEngine {
-  currentDoc: any;
+  currentDoc: Doc;
   constructor(private docService: DocService) {}
 
-  navigate(documentId) {
+  navigate(documentId: string) {
     this.docService.getDoc(documentId).subscribe(
       doc => this.currentDoc = doc
     );
