+++ angular.io/src/app/nav-engine/sitemap.service.spec.ts
@@ -1,8 +1,32 @@
-import { DocMetadata } from '../model';
+import { fakeAsync, tick } from '@angular/core/testing';
+import { DocMetadata } from './doc.model';
 import { SiteMapService } from './sitemap.service';
 
-xdescribe('SiteMapService', () => {
-  it('can create service', () => {
-    expect(true).toBe(true);
+describe('SiteMapService', () => {
+  let siteMapService: SiteMapService;
+
+  beforeEach(() => {
+    siteMapService = new SiteMapService();
   });
+
+  it('should get News metadata', fakeAsync(() => {
+    siteMapService.getDocMetadata('news').subscribe(
+      metadata => expect(metadata.url).toBe('assets/documents/news.html')
+    );
+    tick();
+  }));
+
+  it('should calculate expected doc url for unknown id', fakeAsync(() => {
+    siteMapService.getDocMetadata('fizbuz').subscribe(
+      metadata => expect(metadata.url).toBe('assets/documents/fizbuz.html')
+    );
+    tick();
+  }));
+
+  it('should calculate expected index doc url for unknown id ending in /', fakeAsync(() => {
+    siteMapService.getDocMetadata('fizbuz/').subscribe(
+      metadata => expect(metadata.url).toBe('assets/documents/fizbuz/index.html')
+    );
+    tick();
+  }));
 });
