+++ angular.io/src/app/nav-engine/sitemap.service.ts
@@ -1,9 +1,11 @@
 import { Injectable } from '@angular/core';
+
+import { Observable } from 'rxjs/Observable';
 import { of } from 'rxjs/observable/of';
 import { BehaviorSubject } from 'rxjs/BehaviorSubject';
 import 'rxjs/add/operator/map';
 
-import { DocMetadata } from '../model';
+import { DocMetadata } from './doc.model';
 
 const siteMap: DocMetadata[] = [
   { 'title': 'Home', 'url': 'assets/documents/home.html', id: 'home'},
@@ -18,20 +20,19 @@ export class SiteMapService {
   getDocMetadata(id: string) {
     const missing = () => this.getMissingMetadata(id);
     return this.siteMap
-        .map(map =>
-         map.find(d => d.id === id) || missing());
+      .map(map =>
+        map.find(d => d.id === id) || missing());
   }
 
-  getMissingMetadata(id: string) {
+  // Alternative way to calculate metadata. Will it be used?
+  private getMissingMetadata(id: string) {
 
-    const fid = id.startsWith('/') ? id.substring(1) : id; // strip leading '/'
+    const filename = id.startsWith('/') ? id.substring(1) : id; // strip leading '/'
 
     return {
       id,
       title: id,
-      url: `assets/documents/${fid}${fid.endsWith('/') ? 'index' : ''}.html`
+      url: `assets/documents/${filename}${filename.endsWith('/') ? 'index' : ''}.html`
     } as DocMetadata;
   }
 }
-
-// Alternative way to calculate metadata
